+++ libsq/core/execz/shellexec.go
@@ -1,4 +1,4 @@
-package cli
+package execz
 
 import (
 	"bytes"
